refactor: 커스텀 훅 정리
