+++ src/model/TeckStack.ts
@@ -1,5 +1,5 @@
 export type TechStackType = {
-	id: string;
-	name: string;
+	stackId: string;
+	stackName: string;
 	stackImg: string;
 };
