+++ src/pages/Teammate/TeammateList.tsx
@@ -1,30 +1,8 @@
-import useTeammate from "../../hooks/useTeammate";
-import { useTeammateStore } from "../../store/teammateStore.ts";
 import PostCard from "../../components/PostCard/PostCard.tsx";
-import { useNavigate } from "react-router-dom";
-import useRepostTeammate from "../../hooks/useRepostTeammate";
+import useTeammateList from "../../hooks/useTeammateList.ts";
 
 const TeammateList = () => {
-  const nav = useNavigate();
-  const { teammates } = useTeammateStore();
-  useTeammate();
-
-  const { repostTeammate } = useRepostTeammate();
-
-  const handleClick = (id: string) => {
-    nav(`/findteammate/${id}`);
-  };
-
-  const handleRepost = () => {
-    if (teammates.length === 0) {
-      alert("재게시할 프로필이 없습니다.");
-      return;
-    }
-
-    const postId = teammates[0].id;
-    repostTeammate(postId);
-  };
-
+  const { handleRepost, teammates, handleClick } = useTeammateList();
   return (
     <>
       <div className="w-[1140px] flex justify-end mt-10">
